fix: fix mapping ipv6-based port mappings
