+++ src/index.ts
@@ -1,5 +1,7 @@
 import childProcess from 'child_process'
 import yaml from 'yaml'
+import mapPorts from './port-mapper'
+
 export interface IDockerComposeOptions {
   cwd?: string
   executablePath?: string
@@ -73,33 +75,6 @@ export type DockerComposePsResult = {
       exposed: { port: number; protocol: string }
     }>
   }>
-}
-
-export const mapPorts = (
-  ports: string
-): Array<{
-  mapped?: { address: string; port: number }
-  exposed: { port: number; protocol: string }
-}> => {
-  const result = !ports
-    ? []
-    : (() => {
-        return ports.split(',').map((untypedPort) => {
-          const exposedFragments = untypedPort.trim().split('->')
-          const [port, protocol] =
-            exposedFragments.length === 1
-              ? exposedFragments[0].split('/')
-              : exposedFragments[1].split('/')
-          const [address, mappedPort] =
-            exposedFragments.length === 2 ? exposedFragments[0].split(':') : []
-          return {
-            exposed: { port: Number(port), protocol },
-            ...(address &&
-              mappedPort && { mapped: { port: Number(mappedPort), address } })
-          }
-        })
-      })()
-  return result
 }
 
 export const mapPsOutput = (output: string): DockerComposePsResult => {
@@ -514,3 +489,5 @@ export const version = async function (
     return Promise.reject(error)
   }
 }
+
+export { mapPorts }
